+++ bake/utopia/project/readme/update.rb
@@ -13,7 +13,7 @@ def update(path: "readme.md", documentation_url: nil)
 	project = Utopia::Project::Base.new(context.root)
 	readme = project.readme_document
 	
-	documentation_url ||= public_documentation_url
+	documentation_url ||= public_documentation_url(project)
 	
 	readme.replace_section("Usage") do |header|
 		current = header
@@ -42,25 +42,15 @@ private
 
 Scope = Struct.new(:documentation_url, :project)
 
-def gemspec_path
-	Dir.glob("*.gemspec", base: context.root).first
-end
-
-def gemspec
-	if gemspec_path = self.gemspec_path
-		@gemspec ||= ::Gem::Specification.load(gemspec_path)
-	end
-end
-
 # The public documentation URL if it can be determined.
-def public_documentation_url
-	if metadata = gemspec.metadata
+def public_documentation_url(project)
+	if metadata = project.gemspec&.metadata
 		if documentation_uri = metadata["documentation_uri"]
 			return documentation_uri
 		end
 	end
 	
-	return gemspec&.homepage
+	return project.gemspec&.homepage
 end
 
 def usage_section(documentation_url, project)