+++ lib/utopia/project/guide.rb
@@ -40,6 +40,31 @@ module Utopia
 				metadata[:order]
 			end
 			
+			def <=> other
+				if order = self.order
+					if other_order = other.order
+						if order < other_order
+							return -1
+						elsif order > other_order
+							return 1
+						end
+					else
+						# If we have order, but the other doesn't, we come first:
+						return -1
+					end
+				end
+				
+				if name = self.name
+					if other_name = other.name
+						return name <=> other_name
+					else
+						return -1
+					end
+				end
+				
+				return 0
+			end
+			
 			README = "readme.md"
 			
 			# The path to the README file for the guide.