+++ releases.md
@@ -1,5 +1,9 @@
 # Changes
 
+## Unreleased
+
+  - Fix schema for `index.yaml` context file.
+
 ## v0.34.0
 
   - Introduce `bake utopia:project:agent:context:update` command to update the agent context from the guides in the project.